requirements.txt
Flask==3.1.1
keras==2.13.1
matplotlib==3.7.2
numpy==1.24.3
opencv_python==4.9.0.80
pandas==2.0.3
pyzmq==27.0.0
scikit_learn==1.3.0
tensorflow==2.13.0
tensorflow_intel==2.13.0
